backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
sqlalchemy>=2.0.35
openai==1.35.3
pydantic>=2.8.0
python-dotenv==1.0.1
Pillow>=11.0.0
requests